requirements.txt
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
google-genai>=1.67.0
streamlit>=1.32.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
